src/test/java/com/leetcode/bfs/medium/MinimumGeneticMutation_433Test.java
package com.leetcode.bfs.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class MinimumGeneticMutation_433Test {
  private final MinimumGeneticMutation_433 solution = new MinimumGeneticMutation_433();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments("AACCGGTT", "AACCGGTA", new String[] {"AACCGGTA"}),
          arguments("AACCGGTT", "AAACGGTA", new String[] {"AACCGGTA", "AACCGCTA", "AAACGGTA"}),
          arguments("AAAAACCC", "AACCCCCC", new String[] {"AAAACCCC", "AAACCCCC", "AACCCCCC"}));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void minMutation(String start, String end, String[] bank, int expectedResult) {
    int actualResult = solution.minMutation(start, end, bank);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
